+++ CHANGELOG.md
@@ -29,6 +29,7 @@ plus de détails sur http://semver.org/lang/fr/
 	* support d'un `thumbnail`
 * module `block-dashboard` ajouté
 * module `block-dashboard-images` ajouté
+* module `block-dashboard-map` ajouté
 * module `block-dashboard-observations` ajouté
 * module `button` : ajout de la couleur rouge
 * module `breadcrumbs` :
@@ -43,7 +44,7 @@ plus de détails sur http://semver.org/lang/fr/
 * module `error-page` ajouté
 * module `feed` ajouté
 * module `feed-date` ajouté
-* module `feed-item` ajouté
+* module `feed-item` : ajouté + refactoring avec Pug
 * module `footer` :
 	* ajout des icônes de flèche au dernier item des colonnes du menu
 	* ajout de l'icône CC